refactor federal cdcc to include cap
