+++ policyengine_us/parameters/gov/states/id/tax/income/deductions/dependent_care_expenses/cap/amount.yaml
@@ -1,11 +1,10 @@
 description: Idaho caps the household and dependent care expense deduction at this amount.
 values:
-  2004-01-01: .inf
   2023-01-01: 12_000
 metadata:
   unit: currency-USD
   period: year
-  label: Idaho dependent care expense deduction cap
+  label: Idaho dependent care expense deduction cap amount
   reference:
   - title: Idaho Statutes 63-3022D 
     href: https://legislature.idaho.gov/statutesrules/idstat/title63/t63ch30/sect63-3022d/
@@ -15,4 +14,4 @@ metadata:
     href: https://tax.idaho.gov/wp-content/uploads/forms/EIN00046/EIN00046_11-06-2023.pdf#page=2
   # The cap was introduced in 2022
   - title: 2022 Idaho Tax Form 40 Instructions
-    href: https://tax.idaho.gov/wp-content/uploads/forms/EIN00046/EIN00046_11-15-2021.pdf#31
+    href: https://tax.idaho.gov/wp-content/uploads/forms/EIN00046/EIN00046_11-15-2021.pdf#page=31
