+++ policyengine_us/parameters/gov/states/id/tax/income/deductions/dependent_care_expenses/cap/available.yaml
@@ -1,0 +1,18 @@
+description: Idaho caps the household and dependent care expense deduction if this is true.
+values:
+  2004-01-01: false
+  2023-01-01: true
+metadata:
+  unit: bool
+  period: year
+  label: Idaho dependent care expense deduction cap available
+  reference:
+  - title: Idaho Statutes 63-3022D 
+    href: https://legislature.idaho.gov/statutesrules/idstat/title63/t63ch30/sect63-3022d/
+  - title: Idaho Statutes 63-3022D # prior to 2023
+    href: https://law.justia.com/codes/idaho/2022/title-63/chapter-30/section-63-3022d/
+  - title: 2023 Idaho Tax Form 40 Instructions
+    href: https://tax.idaho.gov/wp-content/uploads/forms/EIN00046/EIN00046_11-06-2023.pdf#page=2
+  # The cap was introduced in 2022
+  - title: 2022 Idaho Tax Form 40 Instructions
+    href: https://tax.idaho.gov/wp-content/uploads/forms/EIN00046/EIN00046_11-15-2021.pdf#page=31
